requirements.txt
Django==4.2.9
django-adminlte3==0.1.6
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
Pillow==10.2.0
django-filter==23.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1